Move magit-status to C-c g
C-c <letter> is the official user space for keybindings, and it's
recommended not to bind C-g to anything in prefix maps, as it's the
natural way to abort.

So I'll retrain my muscle memory to avoid ending up in random
magit-status buffers from time to time.

Didn't use counsel-git anyway (use projectile-find-file instead).